Linear Approach
Now the linearApproach does not measure the angle during the drive anymore, because
the angle had to much noise, which came from small vibrations of the camera and resulted in too bad measurements of the angle. Indeed the measurments during drive are so bad,
that you cannot use the results at all.
Thus it should be never tried again to make measurements with an apriltag during drive.
The distance is now calculated with the data from the first meaurement,
which is taken, when the robot is NOT moving. Then the robot turns about epsilon and
drives the distance by using the drive_forward function.
Thus there is no measurement taken during the drive anymore.